+++ src/lib/Components/Bidding/Screens/BidResult.tsx
@@ -29,13 +29,11 @@ interface BidResultProps {
 
 const messageForPollingTimeout = {
   title: "Bid processing",
-  description: `
-    We’re receiving a high volume of traffic
-    and your bid is still processing.
-
-    If you don’t receive an update soon,
-    please contact [support@artsy.net](mailto:support@artsy.net).
-  `,
+  description:
+    "We’re receiving a high volume of traffic\n" +
+    "and your bid is still processing.\n\n" +
+    "If you don’t receive an update soon,\n" +
+    "please contact [support@artsy.net](mailto:support@artsy.net).",
 }
 
 const Icons = {
@@ -80,13 +78,15 @@ export class BidResult extends React.Component<BidResultProps> {
             <Flex alignItems="center">
               <Icon20 source={Icons[status] || require("../../../../../images/circle-x-red.png")} />
 
-              <Title m={4}>
+              <Title mt={0} mb={5}>
                 {status === "PENDING" ? messageForPollingTimeout.title : message_header || "You’re the highest bidder"}
               </Title>
 
-              <Markdown>
-                {status === "PENDING" ? messageForPollingTimeout.description : message_description_md || ""}
-              </Markdown>
+              {status !== "WINNING" && (
+                <Markdown mb={5}>
+                  {status === "PENDING" ? messageForPollingTimeout.description : message_description_md}
+                </Markdown>
+              )}
 
               {this.shouldDisplayTimer(status) && <Timer liveStartsAt={live_start_at} endsAt={end_at} />}
             </Flex>