+++ src/lib/Components/Bidding/Screens/Registration.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { NativeModules, NavigatorIOS, View, ViewProperties } from "react-native"
-import { createFragmentContainer, graphql, RelayPaginationProp } from "react-relay"
+import { commitMutation, createFragmentContainer, graphql, RelayPaginationProp } from "react-relay"
 import styled from "styled-components/native"
 import stripe from "tipsi-stripe"
 
@@ -18,6 +18,7 @@ import { Timer } from "../Components/Timer"
 import { Title } from "../Components/Title"
 import { Address, PaymentCardTextFieldParams, StripeToken } from "../types"
 
+import { Registration_me } from "__generated__/Registration_me.graphql"
 import { Registration_sale } from "__generated__/Registration_sale.graphql"
 
 const Emission = NativeModules.Emission || {}
@@ -26,6 +27,7 @@ stripe.setOptions({ publishableKey: Emission.stripePublishableKey })
 
 export interface RegistrationProps extends ViewProperties {
   sale: Registration_sale
+  me: Registration_me
   relay?: RelayPaginationProp
   navigator?: NavigatorIOS
 }
@@ -36,15 +38,59 @@ interface RegistrationState {
   creditCardToken?: StripeToken
   conditionsOfSaleChecked: boolean
   isLoading: boolean
+  requiresPaymentInformation: boolean
 }
 
+const creditCardMutation = graphql`
+  mutation RegistrationCreateCreditCardMutation($input: CreditCardInput!) {
+    createCreditCard(input: $input) {
+      credit_card {
+        id
+        brand
+        name
+        last_digits
+        expiration_month
+        expiration_year
+      }
+    }
+  }
+`
+const bidderMutation = graphql`
+  mutation RegistrationCreateBidderMutation($input: CreateBidderInput!) {
+    createBidder(input: $input) {
+      bidder {
+        id
+        qualified_for_bidding
+      }
+    }
+  }
+`
+
 export class Registration extends React.Component<RegistrationProps, RegistrationState> {
-  state = {
-    billingAddress: null,
-    creditCardToken: null,
-    creditCardFormParams: null,
-    conditionsOfSaleChecked: false,
-    isLoading: false,
+  constructor(props) {
+    super(props)
+
+    const { has_credit_cards } = this.props.me
+    const requiresPaymentInformation = !has_credit_cards
+
+    this.state = {
+      billingAddress: null,
+      creditCardToken: null,
+      creditCardFormParams: null,
+      conditionsOfSaleChecked: false,
+      requiresPaymentInformation,
+      isLoading: false,
+    }
+  }
+
+  canCreateBidder() {
+    const { billingAddress, creditCardToken, conditionsOfSaleChecked } = this.state
+
+    if (this.state.requiresPaymentInformation) {
+      return billingAddress && creditCardToken && conditionsOfSaleChecked
+    } else {
+      return conditionsOfSaleChecked
+    }
   }
 
   onPressConditionsOfSale = () => {
@@ -63,6 +109,51 @@ export class Registration extends React.Component<RegistrationProps, Registratio
     this.setState({ conditionsOfSaleChecked: !this.state.conditionsOfSaleChecked })
   }
 
+  register() {
+    this.setState({ isLoading: true })
+
+    this.state.requiresPaymentInformation ? this.createCreditCardAndBidder() : this.createBidder()
+  }
+
+  async createCreditCardAndBidder() {
+    const { billingAddress, creditCardFormParams } = this.state
+    const token = await stripe.createTokenWithCard({
+      ...creditCardFormParams,
+      name: billingAddress.fullName,
+      addressLine1: billingAddress.addressLine1,
+      addressLine2: billingAddress.addressLine2,
+      addressCity: billingAddress.city,
+      addressState: billingAddress.state,
+      addressZip: billingAddress.postalCode,
+    })
+
+    commitMutation(this.props.relay.environment, {
+      onCompleted: () => this.createBidder(),
+      onError: () => {
+        console.log("ERROR!")
+      },
+      mutation: creditCardMutation,
+      variables: {
+        input: {
+          token: token.tokenId,
+        },
+      },
+    })
+  }
+
+  createBidder() {
+    commitMutation(this.props.relay.environment, {
+      onCompleted: () => {
+        console.log("COMPLETED!")
+      },
+      onError: () => {
+        console.log("ERROR!")
+      },
+      mutation: bidderMutation,
+      variables: { input: { sale_id: this.props.sale.id } },
+    })
+  }
+
   render() {
     const { live_start_at, end_at, is_preview, start_at } = this.props.sale
 
@@ -78,14 +169,16 @@ export class Registration extends React.Component<RegistrationProps, Registratio
               </SerifSemibold18>
             </Flex>
 
-            <PaymentInfo
-              navigator={this.props.navigator}
-              onCreditCardAdded={this.onCreditCardAdded.bind(this)}
-              onBillingAddressAdded={this.onBillingAddressAdded.bind(this)}
-              billingAddress={this.state.billingAddress}
-              creditCardFormParams={this.state.creditCardFormParams}
-              creditCardToken={this.state.creditCardToken}
-            />
+            {this.state.requiresPaymentInformation && (
+              <PaymentInfo
+                navigator={this.props.navigator}
+                onCreditCardAdded={this.onCreditCardAdded.bind(this)}
+                onBillingAddressAdded={this.onBillingAddressAdded.bind(this)}
+                billingAddress={this.state.billingAddress}
+                creditCardFormParams={this.state.creditCardFormParams}
+                creditCardToken={this.state.creditCardToken}
+              />
+            )}
           </View>
 
           <View>
@@ -100,7 +193,8 @@ export class Registration extends React.Component<RegistrationProps, Registratio
                 text="Complete Registration"
                 inProgress={this.state.isLoading}
                 selected={this.state.isLoading}
-                onPress={() => null}
+                onPress={this.canCreateBidder() ? () => this.register() : null}
+                disabled={!this.canCreateBidder()}
               />
             </Flex>
           </View>
@@ -114,9 +208,8 @@ const LinkText = styled.Text`
   text-decoration-line: underline;
 `
 
-export const RegistrationScreen = createFragmentContainer(
-  Registration,
-  graphql`
+export const RegistrationScreen = createFragmentContainer(Registration, {
+  sale: graphql`
     fragment Registration_sale on Sale {
       id
       end_at
@@ -125,5 +218,10 @@ export const RegistrationScreen = createFragmentContainer(
       name
       start_at
     }
-  `
-)
+  `,
+  me: graphql`
+    fragment Registration_me on Me {
+      has_credit_cards
+    }
+  `,
+})